+++ src/test/java/com/training/salestaxes/calculation/RoundingStrategyTest.java
@@ -1,5 +1,7 @@
 package com.training.salestaxes.calculation;
 
+import java.math.BigDecimal;
+
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Before;
@@ -19,7 +21,7 @@ public class RoundingStrategyTest
 	@Test
 	public void shouldRoundToNearestHalf_Up()
 	{
-		double roundedValue = roundingStrategy.roundUp(14.99d, 10);
+		double roundedValue = roundingStrategy.roundUp(BigDecimal.valueOf(14.99), 10);
 
 		assertEquals(1.5d, roundedValue, 0);
 	}
@@ -27,7 +29,7 @@ public class RoundingStrategyTest
 	@Test
 	public void shouldRoundToNearestHalf_Down()
 	{
-		double roundedValue = roundingStrategy.roundUp(14.21d, 10);
+		double roundedValue = roundingStrategy.roundUp(BigDecimal.valueOf(14.21), 10);
 
 		assertEquals(1.45d, roundedValue, 0);
 	}
